move suffixed things into separate file
